# backend/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "description": "註冊新用戶賬號",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "註冊成功"
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "409": {
            "description": "用戶已存在"
          }
        },
        "summary": "用戶註冊",
        "tags": [
          "認證系統"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "用戶登入系統",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登入成功"
          },
          "401": {
            "description": "用戶名或密碼錯誤"
          }
        },
        "summary": "用戶登入",
        "tags": [
          "認證系統"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "description": "發送密碼重置郵件",
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密碼重置郵件已發送"
          },
          "404": {
            "description": "用戶不存在"
          }
        },
        "summary": "忘記密碼",
        "tags": [
          "認證系統"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "description": "使用重置令牌重置密碼",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密碼重置成功"
          },
          "400": {
            "description": "重置令牌無效或已過期"
          }
        },
        "summary": "重置密碼",
        "tags": [
          "認證系統"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "用戶登出系統",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "登出成功"
          },
          "401": {
            "description": "未認證用戶"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "用戶登出",
        "tags": [
          "認證系統"
        ]
      }
    },
    "/api/v1/auth/profile": {
      "get": {
        "description": "獲取當前用戶的個人資料",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "獲取個人資料成功"
          },
          "401": {
            "description": "未認證用戶"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "獲取個人資料",
        "tags": [
          "認證系統"
        ]
      }
    },
    "/api/v1/buyers/me/cart": {
      "get": {
        "description": "獲取當前用戶的購物車內容",
        "operationId": "CartController_getCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": "獲取購物車成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "獲取購物車",
        "tags": [
          "購物車管理"
        ]
      }
    },
    "/api/v1/buyers/me/cart/items": {
      "post": {
        "description": "將商品添加到購物車中",
        "operationId": "CartController_addItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "商品已加入購物車"
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "商品不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "添加商品到購物車",
        "tags": [
          "購物車管理"
        ]
      }
    },
    "/api/v1/buyers/me/cart/items/{itemId}": {
      "patch": {
        "description": "更新購物車中指定項目的數量或其他屬性",
        "operationId": "CartController_updateItem",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "購物車項目ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "購物車項目已更新"
          },
          "400": {
            "description": "請求參數錯誤"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "購物車項目不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新購物車項目",
        "tags": [
          "購物車管理"
        ]
      },
      "delete": {
        "description": "從購物車中移除指定項目",
        "operationId": "CartController_removeItem",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "description": "購物車項目ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "項目已移除"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "購物車項目不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "移除購物車項目",
        "tags": [
          "購物車管理"
        ]
      }
    },
    "/api/v1/buyers/me/checkout": {
      "post": {
        "description": "將購物車中的商品進行結帳，創建訂單",
        "operationId": "OrderController_checkout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "結帳成功"
          },
          "400": {
            "description": "請求參數錯誤或購物車為空"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "購物車項目不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "結帳",
        "tags": [
          "買家訂單管理"
        ]
      }
    },
    "/api/v1/buyers/me/orders": {
      "get": {
        "description": "獲取當前用戶的所有訂單列表",
        "operationId": "OrderController_getOrders",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每頁數量",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "頁碼",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取訂單列表成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "獲取訂單列表",
        "tags": [
          "買家訂單管理"
        ]
      }
    },
    "/api/v1/buyers/me/orders/{orderId}": {
      "get": {
        "description": "獲取指定訂單的詳細資訊",
        "operationId": "OrderController_getOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "description": "訂單ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "獲取訂單詳情成功"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "訂單不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "獲取訂單詳情",
        "tags": [
          "買家訂單管理"
        ]
      }
    },
    "/api/v1/buyers/me/orders/{orderId}/cancel": {
      "patch": {
        "description": "取消指定的訂單（僅限未處理狀態）",
        "operationId": "OrderController_cancelOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "description": "訂單ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "訂單已取消"
          },
          "400": {
            "description": "訂單狀態不允許取消"
          },
          "401": {
            "description": "未認證用戶"
          },
          "403": {
            "description": "權限不足"
          },
          "404": {
            "description": "訂單不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "取消訂單",
        "tags": [
          "買家訂單管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zipperoo 電商平台 API",
    "description": "Zipperoo 電商平台後端 API 文檔",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string",
            "description": "用戶帳號",
            "example": "john_doe",
            "minLength": 3
          },
          "password": {
            "type": "string",
            "description": "用戶密碼",
            "example": "password123",
            "minLength": 6
          },
          "username": {
            "type": "string",
            "description": "用戶名稱",
            "example": "John Doe",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "description": "電子郵件地址",
            "example": "john@example.com"
          },
          "phone": {
            "type": "string",
            "description": "手機號碼",
            "example": "+886912345678"
          },
          "role": {
            "type": "string",
            "description": "用戶角色",
            "enum": [
              "BUYER",
              "SELLER"
            ],
            "example": "BUYER"
          },
          "shopName": {
            "type": "string",
            "description": "商店名稱（僅限賣家）",
            "example": "John's Shop"
          },
          "description": {
            "type": "string",
            "description": "商店描述（僅限賣家）",
            "example": "專營電子產品的優質商店"
          }
        },
        "required": [
          "account",
          "password",
          "username",
          "role"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string",
            "description": "用戶帳號",
            "example": "john_doe",
            "minLength": 3
          },
          "password": {
            "type": "string",
            "description": "用戶密碼",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "account",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
      "AddCartItemDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {}
      },
      "CheckoutDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}